requirements.txt
PyPDF2
python-docx
pandas
scikit-learn
